ทำ controller ของ Results - close #132

+++ backend/controllers/results_controller.go
@@ -23,10 +23,9 @@ type ResultsController struct {
 
 // Results struct eiei
 type Results struct {
-	Grade     float
+	Grade     float64
 	StudentID int
 	YearID    int
-	//termID   	int
 	SubjectID int
 }
